Began implementing an integration test for FcgiServerInterface with NGINX.

+++ fcgi/test/test/curl_easy_handle_classes_test.sh
@@ -16,18 +16,25 @@ cp --update ${TEST_SRCDIR}/__main__/fcgi/test/test/curl_easy_handle_classes_inte
 # text which contains replacement patterns which determine which environment
 # variables will be substituted in STDIN.
 envsubst '${TEST_TMPDIR}' < ${TEST_SRCDIR}/__main__/fcgi/test/test/curl_easy_handle_classes_test_nginx.conf.template \
-         > ${TEST_TMPDIR}/curl_easy_handle_classes_test_nginx.conf &&
-nginx -c ${TEST_TMPDIR}/curl_easy_handle_classes_test_nginx.conf \
-      -g "daemon off;" &
+         > ${TEST_TMPDIR}/curl_easy_handle_classes_test_nginx.conf
 
-# Waits so that NGINX has time to initialize.
-sleep 0.1 &&
-# Executes the internal test which uses libcurl. The test will fail if the
-# NGINX instance was not created or is not ready.
-${TEST_SRCDIR}/__main__/fcgi/test/test/curl_easy_handle_classes_internal_test
-test_status=$?
-
-# kill is needed because the command "nginx -s quit" assumes that the default
-# location for the process ID of the NGINX master process is used.
-kill -QUIT $(cat ${TEST_TMPDIR}/nginx.pid)
-exit $test_status
+if [[ $? -eq 0 ]]; then
+    nginx -c ${TEST_TMPDIR}/curl_easy_handle_classes_test_nginx.conf \
+          -g "daemon off;" &
+    if [[ $? -eq 0 ]]; then
+        # Waits so that NGINX has time to initialize.
+        sleep 0.1 &&
+        # Executes the internal test which uses libcurl. The test will fail if the
+        # NGINX instance was not created or is not ready.
+        ${TEST_SRCDIR}/__main__/fcgi/test/test/curl_easy_handle_classes_internal_test
+        test_status=$?
+        # kill is needed because the command "nginx -s quit" assumes that the default
+        # location for the process ID of the NGINX master process is used.
+        kill -QUIT $(cat ${TEST_TMPDIR}/nginx.pid)
+        exit $test_status
+    else
+        exit 1
+    fi
+else
+    exit 1
+fi
